[svn] Remove warning under AIX.

+++ src/cookies.c
@@ -636,7 +636,7 @@ parse_set_cookies (const char *sc,
   if (!silent)
     logprintf (LOG_NOTQUIET,
 	       _("Syntax error in Set-Cookie: %s at position %d.\n"),
-	       escnonprint (sc), p - sc);
+	       escnonprint (sc), (int) (p - sc));
   return NULL;
 }
